requirements.txt
absl-py==1.4.0
accelerate==0.22.0
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
controlnet-aux==0.0.7
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
diffusers @ git+https://github.com/huggingface/diffusers.git@2ab170499eaaf7adfa24a80e0e2717c916f598f1
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
invisible-watermark==0.2.0
ipykernel==6.25.2
ipyplot==1.1.1
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.11.2
sentencepiece==0.1.99
shortuuid==1.0.11
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.2
sympy==1.12
tifffile==2023.8.30
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
zipp==3.16.2
